backend/ml_service/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
psycopg2-binary
python-dotenv
